Moved integration helpers
